refactor: Organize components and utils into separate files
This commit refactors the project structure for improved readability,
maintainability, and reusability, addressing your feedback.

Key changes:
- Created `src/components` directory.
- Moved UI components (ProblemForm, ProblemCard, TodaysReviewSection,
  ReviewCalendarSection, ActionBar, StatisticsSection) from App.jsx
  into their respective files within `src/components/`.
- Created `src/utils` directory.
- Moved helper functions (getNextReviewDates, getStreak) from App.jsx
  into `src/utils/dateUtils.js` and `src/utils/problemUtils.js`
  respectively.
- Updated all import statements in `src/App.jsx` and within the
  newly created component files to reflect the new file structure.

The application's functionality remains unchanged. A post-refactoring
manual testing checklist has been prepared to ensure everything
works as expected.
